# makes guest API more similar to host