--- a/examples/noop/main.go
+++ b/examples/noop/main.go
@@ -17,13 +17,13 @@
 package main
 
 import (
-	"sigs.k8s.io/kube-scheduler-wasm-extension/guest"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/filter"
 )
 
 func main() {
-	guest.FilterPlugin = api.FilterFunc(filterNoop)
+	filter.Plugin = api.FilterFunc(filterNoop)
 }
 
 // filterNoop doesn't do anything. This is used to test base-case performance.
-func filterNoop(api.NodeInfo, api.Pod) (status *api.Status) { return }
+func filterNoop(api.Pod, api.NodeInfo) (status *api.Status) { return }