--- a/guest/api/types.go
+++ b/guest/api/types.go
@@ -16,29 +16,30 @@
 
 package api
 
-import protoapi "sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto/api"
+import (
+	protoapi "sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto/api"
+	meta "sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto/meta"
+)
 
-// Filter is a WebAssembly implementation of framework.FilterPlugin.
-type Filter interface {
-	Filter(FilterArgs) (statusCode StatusCode, statusReason string)
+// FilterPlugin is a WebAssembly implementation of framework.FilterPlugin.
+type FilterPlugin interface {
+	Filter(nodeInfo NodeInfo, pod Pod) *Status
 }
 
-// FilterFunc adapts an ordinary function to a Filter.
-type FilterFunc func(FilterArgs) (statusCode StatusCode, statusReason string)
+// FilterFunc adapts an ordinary function to a FilterPlugin.
+type FilterFunc func(nodeInfo NodeInfo, pod Pod) *Status
 
 // Filter returns f(a).
-func (f FilterFunc) Filter(a FilterArgs) (statusCode StatusCode, statusReason string) {
-	return f(a)
-}
-
-// FilterArgs are the arguments to a Filter.
-//
-// Note: The arguments are lazy fetched to avoid overhead for properties not in use.
-type FilterArgs interface {
-	NodeInfo() NodeInfo
-	Pod() *protoapi.Pod
+func (f FilterFunc) Filter(nodeInfo NodeInfo, pod Pod) *Status {
+	return f(nodeInfo, pod)
 }
 
 type NodeInfo interface {
 	Node() *protoapi.Node
+}
+
+type Pod interface {
+	Metadata() *meta.ObjectMeta
+	Spec() *protoapi.PodSpec
+	Status() *protoapi.PodStatus
 }