--- a/guest/internal/imports/host.go
+++ b/guest/internal/imports/host.go
@@ -16,10 +16,30 @@
 
 package imports
 
-import "runtime"
+import (
+	"runtime"
 
-// StatusReason overwrites the status reason
-func StatusReason(reason string) {
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
+)
+
+// StatusToCode returns a WebAssembly compatible result for the input status,
+// after sending any reason to the host.
+func StatusToCode(s *api.Status) uint32 {
+	// Nil status is the same as one with a success code.
+	if s == nil || s.Code == api.StatusCodeSuccess {
+		return uint32(api.StatusCodeSuccess)
+	}
+
+	// WebAssembly Core 2.0 (DRAFT) only includes numeric types. Return the
+	// reason using a host function.
+	if reason := s.Reason; reason != "" {
+		statusReason(reason)
+	}
+	return uint32(s.Code)
+}
+
+// statusReason overwrites the status reason
+func statusReason(reason string) {
 	ptr, size := stringToPtr(reason)
 	_statusReason(ptr, size)
 	runtime.KeepAlive(reason) // keep reason alive until ptr is no longer needed.