--- a/internal/e2e/profiler/profiler.go
+++ b/internal/e2e/profiler/profiler.go
@@ -18,12 +18,14 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"io"
 	"log"
 	"os"
 
 	"github.com/stealthrocket/wzprof"
 	"github.com/tetratelabs/wazero/experimental"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/kubernetes/pkg/scheduler/framework"
 
 	wasm "sigs.k8s.io/kube-scheduler-wasm-extension/scheduler/plugin"
@@ -56,7 +58,10 @@ func main() {
 	)
 
 	// Pass the profiling context to the plugin.
-	plugin, err := wasm.NewFromConfig(ctx, wasm.PluginConfig{GuestPath: guestPath})
+	plugin, err := wasm.New(&runtime.Unknown{
+		ContentType: runtime.ContentTypeJSON,
+		Raw:         []byte(fmt.Sprintf(`{"guestPath": "%s"}`, guestPath)),
+	}, nil)
 	if err != nil {
 		log.Panicln("failed to create plugin:", err)
 	}