--- a/scheduler/plugin/mem.go
+++ b/scheduler/plugin/mem.go
@@ -28,22 +28,29 @@ type valueType interface {
 	MarshalToSizedBuffer(dAtA []byte) (int, error)
 }
 
-func marshalIfUnderLimit(mem wazeroapi.Memory, vt valueType, buf uint32, bufLimit bufLimit) (vLen int) {
+func marshalIfUnderLimit(mem wazeroapi.Memory, vt valueType, buf uint32, bufLimit bufLimit) int {
 	// First, see if the caller passed enough memory to serialize the object.
-	vLen = vt.Size()
-	if vLen > int(bufLimit) {
-		return // caller can retry with a larger limit
-	} else if vLen == 0 {
-		return // nothing to write
+	vLen := vt.Size()
+	if vLen == 0 {
+		return 0 // nothing to write
 	}
 
-	// Write directly to the wasm memory.
-	wasmMem, ok := mem.Read(buf, uint32(vLen))
-	if !ok { // caller passed a length outside memory
-		panic("out of memory")
+	// Next, see if the value will fit inside the buffer.
+	if vLen > int(bufLimit) {
+		// If it doesn't fit, the caller can decide to retry with a larger
+		// buffer or fail.
+		return vLen
 	}
-	if _, err := vt.MarshalToSizedBuffer(wasmMem); err != nil {
-		panic(err)
+
+	// Now, we know the value isn't too large to fit in the buffer. Write it
+	// directly to the Wasm memory.
+	if wasmMem, ok := mem.Read(buf, uint32(vLen)); !ok {
+		panic("out of memory") // Bug: caller passed a length outside memory
+	} else if _, err := vt.MarshalToSizedBuffer(wasmMem); err != nil {
+		panic(err) // Bug: in marshaller.
 	}
-	return
+
+	// Success: return the bytes written, so that the caller can unmarshal from
+	// a sized buffer.
+	return vLen
 }