--- a/scheduler/test/testdata/error/not_plugin.wat
+++ b/scheduler/test/testdata/error/not_plugin.wat
@@ -1,0 +1,4 @@
+(module $not_plugin
+  ;; exports a memory but nothing else. This is invalid!
+  (memory (export "memory") 1 1)
+)