… examples/filter-simple/main.wasm
… examples/noop/main.wasm
… scheduler/test/testdata/error/not_plugin.wasm