@codefromthecrypt 2023-06-07
ps I ran tests manually, and they pass. after https://github.com/kubernetes-sigs/kube-scheduler-wasm-extension/pull/30 we don't need manual verification