@codefromthecrypt 2023-06-07
I should disclose the simple wasm filter grew 9KB over this, but it is 9/876 and probably unanimously worth it