@codefromthecrypt 2023-06-08
@anuraaga can you take another look and approve or let me know if any glitches left? since this has no formal approvals to lose yet, I snuck in some more godoc and polish.